+++ packages/dialtone-vue2/recipes/conversation_view/feed_item_row/feed_item_row.vue
@@ -385,13 +385,13 @@ export default {
     right: var(--dt-space-450);
   }
 
-  &:deep(.dt-item-layout) {
+  &:deep(> .dt-item-layout) {
     font: var(--dt-typography-body-compact-base);
     min-height: initial;
     padding: 0px;
   }
 
-  &:deep(.dt-item-layout--left) {
+  &:deep(> .dt-item-layout > .dt-item-layout--left) {
     align-self: flex-start;
     text-align: end;
     display: block;
@@ -401,12 +401,12 @@ export default {
     min-width: calc(var(--dt-space-600) + var(--dt-space-300));
   }
 
-  &:deep(.dt-item-layout--right) {
+  &:deep(> .dt-item-layout > .dt-item-layout--right) {
     padding: 0;
     min-width: initial;
   }
 
-  &:deep(.dt-item-layout--bottom) {
+  &:deep(> .dt-item-layout > .dt-item-layout--bottom) {
     display: flex;
     flex-direction: column;
     margin-top: 0;