+++ packages/dialtone-vue3/recipes/conversation_view/editor/editor.test.js
@@ -17,6 +17,7 @@ let alignJustifyBtn;
 let orderedListBtn;
 let blockquoteBtn;
 let codeblockBtn;
+let quickRepliesBtn;
 let addLinkBtn;
 
 const testText = 'In the beginning, it was a nice day.';
@@ -56,6 +57,7 @@ const _setChildWrappers = () => {
   orderedListBtn = wrapper.find('[data-qa="dt-editor-ordered-list-items-btn"]');
   blockquoteBtn = wrapper.find('[data-qa="dt-editor-blockquote-btn"]');
   codeblockBtn = wrapper.find('[data-qa="dt-editor-code-block-btn"]');
+  quickRepliesBtn = wrapper.find('[data-qa="dt-editor-quick-replies-btn"]');
 };
 
 const _mountWrapper = () => {
@@ -144,6 +146,10 @@ describe('DtRecipeEditor tests', () => {
       expect(blockquoteBtn.exists()).toBe(true);
     });
 
+    it('should contain quick replies button', function () {
+      expect(quickRepliesBtn.exists()).toBe(true);
+    });
+
     it('should contain code block button', function () {
       expect(codeblockBtn.exists()).toBe(true);
     });
@@ -175,6 +181,10 @@ describe('DtRecipeEditor tests', () => {
 
         expect(wrapper
           .find('[data-qa="dt-editor-list-items-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
           .exists()).toBe(true);
 
         expect(wrapper
@@ -213,6 +223,10 @@ describe('DtRecipeEditor tests', () => {
           .exists()).toBe(true);
 
         expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
           .find('[data-qa="dt-editor-add-link-btn"]')
           .exists()).toBe(true);
       });
@@ -245,6 +259,10 @@ describe('DtRecipeEditor tests', () => {
 
         expect(wrapper
           .find('[data-qa="dt-editor-list-items-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
           .exists()).toBe(true);
 
         expect(wrapper
@@ -283,6 +301,10 @@ describe('DtRecipeEditor tests', () => {
           .exists()).toBe(true);
 
         expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
           .find('[data-qa="dt-editor-add-link-btn"]')
           .exists()).toBe(true);
       });
@@ -318,6 +340,49 @@ describe('DtRecipeEditor tests', () => {
           .exists()).toBe(false);
 
         expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-add-link-btn"]')
+          .exists()).toBe(true);
+      });
+    });
+
+    describe('When quick replies button is disabled', () => {
+      beforeEach(async () => {
+        _mountWrapper();
+        await wrapper.setProps({ showQuickRepliesButton: false });
+        await wrapper.vm.$nextTick();
+        _setChildWrappers();
+      });
+
+      it('should not appear in the formatting options', () => {
+        expect(wrapper
+          .find('[data-qa="dt-editor-bold-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-italics-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-underline-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-strike-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-list-items-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
+          .exists()).toBe(false);
+
+        expect(wrapper
           .find('[data-qa="dt-editor-add-link-btn"]')
           .exists()).toBe(true);
       });
@@ -350,6 +415,10 @@ describe('DtRecipeEditor tests', () => {
 
         expect(wrapper
           .find('[data-qa="dt-editor-list-items-btn"]')
+          .exists()).toBe(true);
+
+        expect(wrapper
+          .find('[data-qa="dt-editor-quick-replies-btn"]')
           .exists()).toBe(true);
 
         expect(wrapper
@@ -466,6 +535,14 @@ describe('DtRecipeEditor tests', () => {
         const editorHtmlOutput = editor.html().replaceAll(/[\n\r]/g, '').replaceAll(' ', '');
         expect(editorHtmlOutput)
           .toContain(expectedHtmlOutput.replaceAll(' ', ''));
+      });
+    });
+
+    describe('When quick replies button is clicked', () => {
+      it('quick replies clicked event should be fired', async () => {
+        await quickRepliesBtn.trigger('click');
+        await wrapper.vm.$nextTick();
+        expect('quick-replies-click' in wrapper.emitted()).toBeTruthy();
       });
     });
   });