+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+# [9.18.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.17.0...dialtone/v9.18.0) (2024-03-07)
+
+
+### Bug Fixes
+
+* message input issues ([#190](https://github.com/dialpad/dialtone/issues/190)) ([caa07b6](https://github.com/dialpad/dialtone/commit/caa07b6e9e167094a5e91f5cd8499201194823f0))
+* remove markRaw ([#191](https://github.com/dialpad/dialtone/issues/191)) ([74884e2](https://github.com/dialpad/dialtone/commit/74884e2284bdeeb0e824a58b44377e6248d27fdc))
+
+
+### Documentation
+
+* use CodeExampleTabs from badge to button-group ([#185](https://github.com/dialpad/dialtone/issues/185)) ([808015b](https://github.com/dialpad/dialtone/commit/808015b5662379695f7c8b1ea9f787357d8c4c22))
+
+
+### Features
+
+* **Editor:** add quick replies icon to editor ([#193](https://github.com/dialpad/dialtone/issues/193)) ([7a6829f](https://github.com/dialpad/dialtone/commit/7a6829faab0d29498d2184b486c49efc03e7d8bb))
+
+# [9.17.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.16.0...dialtone/v9.17.0) (2024-03-07)
+
+
+### Features
+
+* dialtone stylelint plugin - no mixins rule ([#186](https://github.com/dialpad/dialtone/issues/186)) ([8fe0ffb](https://github.com/dialpad/dialtone/commit/8fe0ffb7f4dd81cbf19e6cbbe37b16ca609973cd))
+
+# [9.16.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.15.2...dialtone/v9.16.0) (2024-03-06)
+
+
+### Bug Fixes
+
+* lower aggression of feed-item-row deep selector ([#189](https://github.com/dialpad/dialtone/issues/189)) ([de20342](https://github.com/dialpad/dialtone/commit/de20342dc0e831f64c704d84786e293423fbb02d))
+
+
+### Features
+
+* **Emoji Row, Time Pill:** feed time pill and emoji style updates ([#187](https://github.com/dialpad/dialtone/issues/187)) ([cd39cbb](https://github.com/dialpad/dialtone/commit/cd39cbb26db054810ee2b8ed42b610157ba90ca1))
+
 ## [9.15.2](https://github.com/dialpad/dialtone/compare/dialtone/v9.15.1...dialtone/v9.15.2) (2024-03-05)